This fixed and implemented the correct IDisposable pattern where needed.

+++ Neo4j.Driver/Neo4j.Driver.Tests.Integration/Reactive/SummaryIT.cs
@@ -35,6 +35,10 @@ namespace Neo4j.Driver.IntegrationTests.Reactive
     {
         public abstract class Specs : AbstractRxIT
         {
+            private bool _disposed = false;
+
+            ~Specs() => Dispose(false);
+
             protected Specs(ITestOutputHelper output, StandAloneIntegrationTestFixture standAlone)
                 : base(output, standAlone)
             {
@@ -218,34 +222,46 @@ namespace Neo4j.Driver.IntegrationTests.Reactive
                     );
             }
 
-            public override void Dispose()
+            protected override void Dispose(bool disposing)
             {
-                using (var session = Server.Driver.Session())
-                {
-                    foreach (var drop in session.Run("CALL db.constraints()").ToList())
-                    {
-                        if (drop.Values.TryGetValue("name", out var name))
-                        {
-                            session.Run($"DROP CONSTRAINT {name}").Consume();
-                        }
-                    }
+                if (_disposed)
+                    return;
 
-                    foreach (var drop in session.Run("CALL db.indexes()").ToList())
+                if (disposing)
+                {
+                    using (var session = Server.Driver.Session())
                     {
-                        if (drop.Values.TryGetValue("name", out var name))
+                        foreach (var drop in session.Run("CALL db.constraints()").ToList())
                         {
-                            session.Run($"DROP INDEX {name}").Consume();
+                            if (drop.Values.TryGetValue("name", out var name))
+                            {
+                                session.Run($"DROP CONSTRAINT {name}").Consume();
+                            }
+                        }
+
+                        foreach (var drop in session.Run("CALL db.indexes()").ToList())
+                        {
+                            if (drop.Values.TryGetValue("name", out var name))
+                            {
+                                session.Run($"DROP INDEX {name}").Consume();
+                            }
                         }
                     }
                 }
 
-                base.Dispose();
+                //Mark as disposed
+                _disposed = true;
+
+                base.Dispose(disposing);
             }
         }
 
         public class Session : Specs
         {
+            private bool _disposed = false;
             private readonly IRxSession rxSession;
+
+            ~Session() => Dispose(false);
 
             public Session(ITestOutputHelper output, StandAloneIntegrationTestFixture standAlone)
                 : base(output, standAlone)
@@ -258,18 +274,31 @@ namespace Neo4j.Driver.IntegrationTests.Reactive
                 return rxSession;
             }
 
-            public override void Dispose()
+            protected override void Dispose(bool disposing)
             {
-                rxSession.Close<int>().WaitForCompletion();
+                if (_disposed)
+                    return;
 
-                base.Dispose();
+                if (disposing)
+                {
+                    rxSession.Close<int>().WaitForCompletion();
+                }
+
+                //Mark as disposed
+                _disposed = true;
+
+                base.Dispose(disposing);
             }
+
         }
 
         public class Transaction : Specs
         {
+            private bool _disposed = false;
             private readonly IRxSession rxSession;
             private readonly IRxTransaction rxTransaction;
+
+            ~Transaction() => Dispose(false);
 
             public Transaction(ITestOutputHelper output, StandAloneIntegrationTestFixture standAlone)
                 : base(output, standAlone)
@@ -283,12 +312,21 @@ namespace Neo4j.Driver.IntegrationTests.Reactive
                 return rxTransaction;
             }
 
-            public override void Dispose()
+            protected override void Dispose(bool disposing)
             {
-                rxTransaction.Commit<int>().WaitForCompletion();
-                rxSession.Close<int>().WaitForCompletion();
+                if (_disposed)
+                    return;
 
-                base.Dispose();
+                if (disposing)
+                {
+                    rxTransaction.Commit<int>().WaitForCompletion();
+                    rxSession.Close<int>().WaitForCompletion();
+                }
+
+                //Mark as disposed
+                _disposed = true;
+
+                base.Dispose(disposing);
             }
         }
     }
